Proposal describing a design for security for the UI and Admin Server.
Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>

Makes the login flows more explicit for the UI
logins

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>

Rename doc and add link in the README.md file

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>

feat: add proposal for UI OIDC flow

Signed-off-by: Nic Townsend <nictownsend@uk.ibm.com>

Major refactor of the proposal to simplify following comments.

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>

Updated some sentences to make the meaning clearer.

Added a section on encrypting the channel from the client to the server.
Clarified the use of the Basic authentication schema in the SCRAM scenario.
Removed the incorrect references to the OAuth2 refresh token.

Signed-off-by: Graeme McRobert <mcrobert@uk.ibm.com>